Add files via upload
Interactive dashboard that provide deep insights on Austin Real Estate.

Used a mix of DAX queries and advanced visualizations to bring the data to life. The dashboard provides real-time insights into business performance whether it's running in profit or loss with just a few clicks. 
The visualizations are tailored with conditional formatting to reflect data-driven color coding, making the dashboard not only functional but also visually appealing.